requirements.txt
python-dotenv
google-generativeai
uvicorn
fastapi